src/clusterer-lambda/requirements.txt
sentence-transformers
boto3
pandas
psycopg2-binary
numpy
hf_xet
spacy